+++ lib/typescript/commonjs/core/createMerge.d.ts
@@ -3,5 +3,5 @@ import type { CreateMergeConfig, MergeResult } from '../types';
 export declare function createMerge<TInput, TOutput extends {
     id: string;
     updatedAt?: string | null;
-}>(config: CreateMergeConfig<TInput, TOutput>): (items: TInput[], protectAfterSeq?: number) => MergeResult;
+}>(config: CreateMergeConfig<TInput, TOutput>): (items: TInput[], snapshotSeq?: number) => MergeResult;
 //# sourceMappingURL=createMerge.d.ts.map